Rename package portmap to iptables
We'll be doing Kubernetes-based port mapping too, so calling this portmap
is too generic.
